enable -Werror at test case

+++ t/adaptive_counting_unittest.cc
@@ -23,7 +23,7 @@ TEST(AdaptiveCounting, Counting) {
         if (i % 50000 == 0) {
             esti = adp_cnt_card(ctx1);
             printf("actual: %lu, estimated: %lu, error: %.2f%%\n", 
-                    i, esti, fabs((double)(esti - i) / i * 100));
+                    (long unsigned int)i, (long unsigned int)esti, fabs((double)(esti - i) / i * 100));
         }
     }
     printf("\n");
@@ -37,7 +37,7 @@ TEST(AdaptiveCounting, Counting) {
         if (i % 50000 == 0) {
             esti = adp_cnt_card_loglog(ctx1);
             printf("actual: %lu, estimated: %lu, error: %.2f%%\n", 
-                    i, esti, fabs((double)(esti - i) / i * 100));
+                    (long unsigned int)i, (long unsigned int)esti, fabs((double)(esti - i) / i * 100));
         }
     }
     printf("\n");
@@ -49,7 +49,7 @@ TEST(AdaptiveCounting, Counting) {
         if (i % 50000 == 0) {
             esti = adp_cnt_card(ctx2);
             printf("actual: %lu, estimated: %lu, error: %.2f%%\n", 
-                    i, esti, fabs((double)(esti - i) / i * 100));
+                    (long unsigned int)i, (long unsigned int)esti, fabs((double)(esti - i) / i * 100));
         }
     }
     printf("\n");
@@ -63,7 +63,7 @@ TEST(AdaptiveCounting, Counting) {
         if (i % 50000 == 0) {
             esti = adp_cnt_card_loglog(ctx2);
             printf("actual: %lu, estimated: %lu, error: %.2f%%\n", 
-                    i, esti, fabs((double)(esti - i) / i * 100));
+                    (long unsigned int)i, (long unsigned int)esti, fabs((double)(esti - i) / i * 100));
         }
     }
     printf("\n");
@@ -106,7 +106,7 @@ TEST(AdaptiveCounting, Merge) {
     adp_cnt_merge_bytes(ctx, buf1, len1, buf2, len2, NULL);
     esti = adp_cnt_card(ctx);
     printf("actual:40000, estimated: %lu, error: %.2f%%\n", 
-            esti, fabs((double)(esti - 40000) / 40000 * 100));
+            (long unsigned int)esti, fabs((double)(esti - 40000) / 40000 * 100));
 
     adp_cnt_fini(tbm2);
     adp_cnt_fini(tbm1);
