+++ t/linear_counting_unittest.cc
@@ -23,7 +23,7 @@ TEST(LinearCounting, Counting) {
         if (i % 50000 == 0) {
             esti = lnr_cnt_card(ctx1);
             printf("actual: %lu, estimated: %lu, error: %.2f%%\n", 
-                    i, esti, fabs((double)(esti - i) / i * 100));
+                    (long unsigned int)i, (long unsigned int)esti, fabs((double)(esti - i) / i * 100));
         }
     }
     printf("\n");
@@ -37,7 +37,7 @@ TEST(LinearCounting, Counting) {
         if (i % 50000 == 0) {
             esti = lnr_cnt_card(ctx2);
             printf("actual: %lu, estimated: %lu, error: %.2f%%\n", 
-                    i, esti, fabs((double)(esti - i) / i * 100));
+                    (long unsigned int)i, (long unsigned int)esti, fabs((double)(esti - i) / i * 100));
         }
     }
     printf("\n");
@@ -82,7 +82,7 @@ TEST(LinearCounting, Merge) {
     lnr_cnt_merge_bytes(ctx, buf1, len1, buf2, len2, NULL);
     esti = lnr_cnt_card(ctx);
     printf("actual:40000, estimated: %lu, error: %.2f%%\n", 
-            esti, fabs((double)(esti - 40000) / 40000 * 100));
+            (long unsigned int)esti, fabs((double)(esti - 40000) / 40000 * 100));
 
     lnr_cnt_fini(tbm2);
     lnr_cnt_fini(tbm1);
